woah, dunno if i'm screwinf everithing up, or this is a great way to make thigs a little more OOB but ithout loosing too much performace. Also very static safe.
TODO: i'm the structre over the stack. This make easy to spot the return
value and prevent evil side-effect (see multiply on quaternionf.h), but
i'm also concerned about polluting stack and when one of the value is
legit to be result without side effect...
Try to build a _safe version where all parameter are given by pointer as
see if this improve something.